codegen: Introduce non-overlapping temporaries
The conflict with setjmp/longjmp originaed from chibicc's push-pop model
for temporaries. Values that linger in the stack when a jump happens may
be unexpectedly clobbered after resuming.

This change introduces a virtual stack device that translates push-pops
to store-loads, each temporary value is given a unique slot in a region
next to the lvar stack, accessed with the base pointer.
Resolves #8.
The stack in memory now looks like this:

(rsp)
tmp_stack <- new
lvar_stack
(rbp)